requirements.txt
joblib==1.0.1
pandas==1.2.4
biopython==1.78
